refactor: refactor message-dialog
