+++ src/app/shared/components/index.ts
@@ -2,5 +2,6 @@ export * from './content-with-side-nav/content-with-side-nav.component';
 export * from './countdown/countdown.component';
 export * from './error-not-found/error-not-found.component';
 export * from './message-dialog/message-dialog.component';
+export * from './message-dialog/message-dialog.model';
 export * from './under-construction/under-construction.component';
 export * from './video-youtube/video-youtube.component';
