+++ src/app/shared/components/message-dialog/message-dialog.model.ts
@@ -1,0 +1,11 @@
+export interface MessageDialogData {
+    title?: string;
+    message: string;
+    type: MessageType;
+}
+
+export enum MessageType {
+    error,
+    warning,
+    information,
+}
